requirements.txt
flask>=2.0.0
h8mail>=2.0.0
pyyaml>=5.1
